+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed volume set size (shrinking) never trigger the actual shrink operation
 - Only call "secondary" after an "mkfs" command if LINSTOR was the one that set the DRBD resource primary
 - Fixed only reset auto-verify-algo, if at least two nodes reported their crypto entries
+- Fixed set-plain-port, set-plain-listen in linstor-config tool using wrong property instances
 
 ## [1.33.2] - 2026-04-08